the Schemas are ordered by categories
